refactor(adapters): extract link generation to links.rs [#305-3/7]
Move normalize_for_pypi, package_to_pypi_link, vulnerability_id_to_link,
and format_package_name from MarkdownFormatter impl in mod.rs into
dedicated links.rs submodule as pub(super) free functions.

Move all link-related unit tests (~257 lines) to links.rs.
Update mod.rs call sites to use links:: prefix.

Closes #325
